refactor(lib): reorganize lib/ by domain + remove cn from lib
- lib/metadata/  : formatDjs, isInterlude, transformMixtapeToMetadata,
                   parseAirTimeDate (+ tests)
- lib/audio/     : constants (reconnect delays, heartbeat…),
                   element (HTML5 audio state codes + defaultState)
- lib/theme/     : daylight (isDayTime, day/night constants + tests),
                   rainbowColors (+ tests)

cn moved back to app/utils/utils.ts directly (not business logic).
Audio constants test reduced to the only meaningful assertion:
RECONNECT_DELAYS must be in strictly ascending order.
