--- app/lib/utils/cn.test.ts
@@ -1,28 +1,0 @@
-import { describe, expect, it } from 'vitest'
-import { cn } from './cn'
-
-describe('cn', () => {
-  it('returns a single class unchanged', () => {
-    expect(cn('foo')).toBe('foo')
-  })
-
-  it('merges multiple classes', () => {
-    expect(cn('foo', 'bar')).toBe('foo bar')
-  })
-
-  it('handles conditional classes', () => {
-    expect(cn('foo', { bar: true, baz: false })).toBe('foo bar')
-  })
-
-  it('resolves tailwind conflicts (last wins)', () => {
-    expect(cn('p-4', 'p-2')).toBe('p-2')
-  })
-
-  it('handles undefined and null gracefully', () => {
-    expect(cn('foo', undefined, null, 'bar')).toBe('foo bar')
-  })
-
-  it('returns empty string for no inputs', () => {
-    expect(cn()).toBe('')
-  })
-})
