+++ app/utils/isInterlude.ts
@@ -1,3 +1,3 @@
-export { interludeWords } from '~/lib/utils/isInterlude'
-import { isInterlude } from '~/lib/utils/isInterlude'
+export { interludeWords } from '~/lib/metadata/isInterlude'
+import { isInterlude } from '~/lib/metadata/isInterlude'
 export default isInterlude
