+++ app/utils/utils.ts
@@ -1,7 +1,11 @@
 import type { Updater } from '@tanstack/vue-table'
+import { type ClassValue, clsx } from 'clsx'
+import { twMerge } from 'tailwind-merge'
 import type { Ref } from 'vue'
 
-export { cn } from '~/lib/utils/cn'
+export function cn(...inputs: ClassValue[]): string {
+  return twMerge(clsx(inputs))
+}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function valueUpdater<T extends Updater<any>>(updaterOrValue: T, ref: Ref) {
